+++ packages/protocol/src/certificate/kinds/Icac.ts
@@ -24,7 +24,6 @@ export class Icac extends OperationalBase<OperationalCertificate.Icac> {
 
     /** Construct the class from an ASN.1/DER encoded certificate */
     static fromAsn1(asn1: Bytes): Icac {
-        Icac.assertDerSize(asn1);
         const cert = Certificate.parseAsn1Certificate(asn1);
         return new Icac(cert as OperationalCertificate.Icac);
     }