+++ packages/protocol/src/certificate/kinds/Noc.ts
@@ -21,7 +21,6 @@ export class Noc extends OperationalBase<OperationalCertificate.Noc> {
 
     /** Construct the class from an ASN.1/DER encoded certificate */
     static fromAsn1(asn1: Bytes) {
-        Noc.assertDerSize(asn1);
         const cert = Certificate.parseAsn1Certificate(asn1);
         return new Noc(cert as OperationalCertificate.Noc);
     }