+++ packages/protocol/src/certificate/kinds/Rcac.ts
@@ -24,7 +24,6 @@ export class Rcac extends OperationalBase<OperationalCertificate.Rcac> {
 
     /** Construct the class from an ASN.1/DER encoded certificate */
     static fromAsn1(asn1: Bytes): Rcac {
-        Rcac.assertDerSize(asn1);
         const cert = Certificate.parseAsn1Certificate(asn1);
         return new Rcac(cert as OperationalCertificate.Rcac);
     }